Add preference for configuring cache location
This allows you to chose from the internal storage and all external storage locations.
Changing the location will not migrate existing cache files, but they will still be
loaded from the old location (and will exentually be pruned).

This also removes the external storage permission for API 19 and above, since
it's no longer required for the app-specific external storage directories.