Improved parsing of carets and parentheses
Reddit markdown like `^(all of this is supposed to be in superscript)` is now being parsed correctly.

There still are differences between this implementation and Reddit's with parentheses within superscript:
In cases like
`^(this is superscript and contains ( opening and closing ) parentheses)`, Reddit's parser ends the superscript with the first occurence of a closing parenthese and displays the second one. This implementation considers balanced parentheses more importantly and displays the first closing parentheses within the superscripted part and uses the second one to end that part.

This behaviour can be changed easily, but it remains questionable to me if Reddit's behaviour should be mimicked