Code now has light-grey background color
See #472

Both inline code (\`) and code blocks (introduced with 4 spaces or a tab)
now have a light-grey background color for better distinction.

Ideally, it would be possible to set that color based on the currently selected
theme, but it looks like the code would have to be refactored for that.

For now, I've chosen \#33333333 because it seems to be working in all themes.