No longer caching thumbnails in memory
The RecyclerView takes care of managing posts on the screen. Whenever a
post is shown its thumbnail is loaded from the file cache. Since that's
done asynchronously there is no performace hit.

Images in this patch are resized by the ImageView. This is not stricly
necesary but makes the display logic simpler and would allow for
the size of thumbnails to be configurable.