Enabling fullscreen video in the internal browser
fixes #504
Thanks to @japanesephundroid for providing a link to the solution.
I ported @cprcrack's work on video enabled WebViews to the
current implementation, the code was taken from
https://github.com/cprcrack/VideoEnabledWebView .
I successfully tested viewing video both normally and in fullscreen
in OxygenOS Beta 25, which runs on Android 8.0.
Due to emulator problems, I wasn't able to test any other version of
Android, so I think this PR has to be tested thoroughly before merging,
especially on custom ROMs and different phone OEMs' implementations.