Added checkbox for disabling "replies to inbox" when replying to comments
When replying to comments, it's now possible to disable that replies will be sent to the inbox. (Potential improvement: preference for default behaviour)

Reddit's API endpoint /sendreplies is used for this.
About the actual implementation, I'm not sure if it would be more elegant to move the actual call into CommentReplyActivity, but then we have to escalate the id of the newly created post, which would entail some refactoring.

On another note, Reddit's API responses surely have to be a joke, right? No sane person could do that...

Fixes #286