Limit the total number of cached flag entries
This is mainly a safeguard, to keep our memory usage constrained, even
if the users sets the entry maxage to 3 months and looks at a LOT of
posts/comments/etc. -- but we guesstimate that nobody will actually
hit that 10k limit (if this should prove to be an issue, the limit
could also be made configurable).