WIP Try to use ActionBar
I moved from ListActivity to BaseActivity to re-use the ActionBar methods,
but now the list is not updated after I added a new blocklist entry. I
couldn't figure out how to update the ListView - I think it has something
to do with the wrapper that's added for the ActionBar layout.