first attempt at creating tables
It has been a while since I used java or any OO Programing language. So apologies for the bad code.
my main issue is that RedReader is not using mono spaced font so the table is bound to be off. 
I just hope that this doesn't end up being more work to fix than write.