Handle non-HTTP URLs using native apps in WebView
Typically, these are deep links to native apps, which can be opened
using intents. This change attempts to open the native app, falling back
to letting the WebViewClient's URL-handling logic take over if no app is
installed that accepts that intent URL. This not only works for
`intent://` URLs, but also specific ones like `twitter://`.

To verify this change, find a link to a Tweet (found some on r/nfl),
open it using the WebView, then click the "Switch to the app" button
Twitter provides. Before this change, you'd get
`net::ERR_UNKNOWN_URL_SCHEME` erorr in the WebView. After the change, if
you don't have a Twitter client installed, you'll still get the same
error, but if you do have a client installed, it'll be opened.

Closes #787